[agent] moved message utils into internal/
